requirements2.txt
asgiref==3.2.3
Babel==2.8.0
backcall==0.1.0
certifi==2019.11.28
chardet==3.0.4
decorator==4.4.1
Django==2.2.13
django-cors-headers==2.1.0
django-extra-views==0.13.0
django-haystack==3.0
django-oscar==2.1
django-oscar-api==2.1.1
django-phonenumber-field==3.0.1
django-tables2==2.2.1
django-treebeard==4.3.1
django-widget-tweaks==1.4.5
djangorestframework==3.11.0
factory-boy==2.12.0
Faker==3.0.0
idna==2.9
ipdb==0.12.3
ipython==7.11.1
ipython-genutils==0.2.0
jedi==0.15.2
parso==0.5.2
pexpect==4.7.0
phonenumbers==8.11.4
phonenumberslite==8.11.4
pickleshare==0.7.5

prompt-toolkit==3.0.2
ptyprocess==0.6.0
purl==1.5
pycountry==19.8.18
Pygments==2.5.2
python-dateutil==2.8.1
pytz==2019.3
razorpay==1.2.0
requests==2.23.0
six==1.13.0
sorl-thumbnail==12.5.0
sqlparse==0.3.0
text-unidecode==1.3
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.8
django-wkhtmltopdf==3.3.0
django-mathfilters==1.0.0
